+++ src/Hacks/ConditionalRetry.cpp
@@ -12,7 +12,7 @@ class $modify (PlayLayer)
     {
         bool v = GameManager::get()->getGameVariable("0026");
 
-        if (PlayLayer::getCurrentPercentInt() > m_level->m_normalPercent.value() && Client::GetModuleEnabled("conditional-retry"))
+        if (PlayLayer::getCurrentPercentInt() > m_level->m_normalPercent.value() && Client::GetModuleEnabled("conditional-retry") && !m_isPracticeMode && !m_player1->m_isPlatformer)
         {
             GameManager::get()->setGameVariable("0026", false);
         }